requirements.txt
numpy
torch
torchvision
opencv-python
matplotlib
pillow
imageio
imageio-ffmpeg
decord
xformers
einops
easydict
tqdm
python-socketio[client]
requests
flask
OpenEXR==3.3.1